Refactoring react for generate infero
