+++ packages/react/src/components/radio-group/radio-group-context.ts
@@ -1,5 +1,5 @@
 import { RadioGroupCore } from '@devextreme/components';
-import React from 'react';
+import { createContext } from 'react';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const RadioGroupContext = React.createContext<RadioGroupCore<any> | undefined>(undefined);
+export const RadioGroupContext = createContext<RadioGroupCore<any> | undefined>(undefined);
