+++ packages/react/src/internal/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './use-second-effect';
 export * from './use-callback-ref';
 export * from './use-required-context';
 export * from './use-core-state';
+export * from './use-id';
